+++ sdk/formrecognizer/ai-form-recognizer/src/formRecognizerClient.ts
@@ -51,21 +51,12 @@ import {
   BeginRecognizeReceiptPoller,
   BeginRecognizeReceiptPollState
 } from "./lro/analyze/receiptPoller";
-import {
-  FormRecognizerRequestBody,
-  RecognizedFormArray,
-  FormPageArray,
-  RecognizedReceiptArray
-} from "./models";
-import {
-  RecognizeContentResultResponse,
-  RecognizeFormResultResponse,
-  RecognizeReceiptResultResponse
-} from "./internalModels";
+import { FormRecognizerRequestBody, RecognizedFormArray, FormPageArray } from "./models";
+import { RecognizeContentResultResponse, RecognizeFormResultResponse } from "./internalModels";
 import {
   toRecognizeFormResultResponse,
   toRecognizeContentResultResponse,
-  toReceiptResultResponse
+  toRecognizeFormResultResponseFromReceipt
 } from "./transforms";
 import { createFormRecognizerAzureKeyCredentialPolicy } from "./azureKeyCredentialPolicy";
 
@@ -106,10 +97,7 @@ export type BeginRecognizeContentOptions = RecognizeContentOptions & {
 /**
  * The Long-Running-Operation (LRO) poller that allows you to wait until form content is recognized.
  */
-export type ContentPollerLike = PollerLike<
-  PollOperationState<FormPageArray>,
-  FormPageArray
->;
+export type ContentPollerLike = PollerLike<PollOperationState<FormPageArray>, FormPageArray>;
 
 /**
  * Options for retrieving recognized content data
@@ -194,8 +182,8 @@ export type BeginRecognizeReceiptsOptions = RecognizeReceiptsOptions & {
  * The Long-Running-Operation (LRO) poller that allows you to wait until receipt(s) are recognized.
  */
 export type ReceiptPollerLike = PollerLike<
-  PollOperationState<RecognizedReceiptArray>,
-  RecognizedReceiptArray
+  PollOperationState<RecognizedFormArray>,
+  RecognizedFormArray
 >;
 
 /**
@@ -696,7 +684,7 @@ export class FormRecognizerClient {
   private async getReceipts(
     resultId: string,
     options?: GetReceiptsOptions
-  ): Promise<RecognizeReceiptResultResponse> {
+  ): Promise<RecognizeFormResultResponse> {
     const realOptions = options || {};
     const { span, updatedOptions: finalOptions } = createSpan(
       "FormRecognizerClient-getRecognizedReceipt",
@@ -708,7 +696,7 @@ export class FormRecognizerClient {
         resultId,
         operationOptionsToRequestOptionsBase(finalOptions)
       );
-      return toReceiptResultResponse(result);
+      return toRecognizeFormResultResponseFromReceipt(result);
     } catch (e) {
       span.setStatus({
         code: CanonicalCode.UNKNOWN,